// src/enum/common.enum.ts
export enum SexEnum {
    MALE = 'Male',
    FEMALE = 'Female'
}

export enum SiteContentEnum {
    PRIVACYPOLICY = 'Privacy Policy',
    TERMCONDITION = 'Terms And Condition',
    ABOUTUS = 'About Us'
}

export enum SortOrderEnum {
    ASC = 'asc',
    DESC = 'desc'
}
export enum OfferTypeEnum {
    PERCENTAGE = 'Percentage',
    FLAT = 'Flat',
    EXCLUSIVE = 'Exclusive'
}
export enum ActiveStatusEnum {
    ACTIVE = 'Active',
    INACTIVE = 'In-active'
}
export enum DealTypeEnum {
    DAILY = 'Daily',
    EXCLUSIVE = 'Exclusive'
}
export enum DealForEnum {
    CURRENT = 'Current',
    UPCOMMING = 'Upcomming'
}