// src/enum/transection.enum.ts
export enum TransectionTypeEnum {
    DEBIT = 'Debit',
    CREDIT = 'Credit'
}

export enum TransectionCategoryTypeEnum {
    WALLET = 'Wallet',
    ORDER = 'Order',
}
export enum TransectionForTypeEnum {
    FORREFERRAL = 'For Referral',
    BEREFERRAL = 'Be Referral',
    SIGNUP = 'SignUp',
    ORDERPAYMENT = 'Order Payment',
    ORDERREFUND = 'Order Refund',
    ADDWALLET = 'Add Wallet',
}
export enum TransectionStatusEnum {
    INITIATED = 'Initiated',
    FAILED = 'Failed',
    SUCCESS = 'Success'
}